# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {},
        "license": {}
    },
    "paths": {
...
        "/ai_arts/api/inferences/DeleteJob": {
            "delete": {
                "description": "delete a inference job",
                "produces": [
                    "application/json"
                ],
                "summary": "Delete a inference job",
                "parameters": [
                    {
                        "type": "string",
                        "description": "inference job Id ",
                        "name": "jobId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/routers.APISuccessResp"
                        }
                    }
                }
            }
        },
...
        "/ai_arts/api/inferences/GetJobDetail": {
            "get": {
                "description": "get inference job detail",
                "produces": [
                    "application/json"
                ],
                "summary": "get inference job detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "inference job Id ",
                        "name": "jobId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/routers.APISuccessResp"
                        }
                    }
                }
            }
        },
        "/ai_arts/api/inferences/GetJobLog": {
            "get": {
                "description": "get inference job log",
                "produces": [
                    "application/json"
                ],
                "summary": "get inference job log",
                "parameters": [
                    {
                        "type": "string",
                        "description": "inference job Id ",
                        "name": "jobId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/routers.APISuccessResp"
                        }
                    }
                }
            }
        },
        "/ai_arts/api/inferences/GetJobStatus": {
            "get": {
                "description": "get inference job status",
                "produces": [
                    "application/json"
                ],
                "summary": "get inference job status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "inference job Id ",
                        "name": "jobId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/routers.APISuccessResp"
                        }
                    }
                }
            }
        },
        "/ai_arts/api/inferences/Infer": {
            "post": {
                "description": "Infer a picture using a running inference job",
                "produces": [
                    "application/json"
                ],
                "summary": "Infer a picture using a running inference job",
                "parameters": [
                    {
                        "type": "string",
                        "description": "inference job Id ",
                        "name": "jobId",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "picture upload to infer",
                        "name": "image",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/routers.APISuccessResp"
                        }
                    }
                }
            }
        },
        "/ai_arts/api/inferences/KillJob": {
            "get": {
                "description": "kill a running inference job",
                "produces": [
                    "application/json"
                ],
                "summary": "kill a running inference job",
                "parameters": [
                    {
                        "type": "string",
                        "description": "inference job Id ",
                        "name": "jobId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/routers.APISuccessResp"
                        }
                    }
                }
            }
        },
        "/ai_arts/api/inferences/ListInferenceJob": {
            "get": {
                "description": "list inference jobs",
                "produces": [
                    "application/json"
                ],
                "summary": "list inference jobs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "which virtual cluster",
                        "name": "vcName",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "page number, from 1",
                        "name": "pageNum",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "count per page",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "search name",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "search status",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/routers.APISuccessResp"
                        }
                    }
                }
            }
        },
        "/ai_arts/api/inferences/PostInferenceJob": {
            "post": {
                "description": "submit a inference job",
                "produces": [
                    "application/json"
                ],
                "summary": "submit a inference job",
                "parameters": [
                    {
                        "description": "json body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.PostInference"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/routers.APISuccessResp"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.PostInference": {
            "type": "object",
            "properties": {
                "desc": {
                    "type": "string"
                },
                "device": {
                    "type": "string"
                },
                "framework": {
                    "type": "string"
                },
                "gpuType": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "jobName": {
                    "type": "string"
                },
                "model_base_path": {
                    "type": "string"
                },
                "resourcegpu": {
                    "type": "integer"
                },
                "userId": {
                    "type": "string"
                },
                "userName": {
                    "type": "string"
                },
                "vcName": {
                    "type": "string"
                }
            }
        },
...
        "routers.APISuccessResp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
    }
}